Allow specifying duration for --ref-rate/-r flag
This commit changes the --ref-rate/-r flag so that it accepts both an
integer in seconds (original behavior) and a custom duration string (new
behavior). This lets the user draw the UI at a higher refresh rate such
as 30fps.

It also makes a minor change in src/tui/proc-table.go, changing from
using `time.After` to `time.Ticker` to make the code slightly more
performant.