feat: SIGKILL procs without kill command afer timeout
We start all procs as CommandContexts instead of Commands, with a
WithCancel context attached. A context cancellation handler which tries
to stop the application gracefully, waits the specified amount of time,
and subsequently SIGKILLs if the process hasn't stopped already is
attached to the CommandContext.

The CommandContext type internally has some synchronization to try to
minimize the chances of races between the context cancellation handler,
and normal termination of applications.