Add `--enable`, `--disable`
Add `--enable` (`$PC_ENABLE`) and `--disable` (`$PC_DISABLE`) arguments
which allow overriding whether or not a process is run.

This can be used to implement optional processes. For example, the
following process will not be run by default:

    processes:
      hello:
          command: echo 'Hello World from Process Compose'
          disabled: true

But a user who wants to enable that functionality can set
`PC_ENABLE=hello` when starting `process-compose` to enable the process.

Related: #177