Add `process-compose project is-ready [--wait]`
`process-compose --detached-with-tui --detach-on-success` is great for
interactive uses, but not so great when scripting `process-compose`.

It's possible to wait for the project to be ready by parsing and polling
the output of `process-compose process list --output=json`, but it's
tedious and requires duplicating the success logic we already have in
`process-compose`.

Therefore, this change adds a new command, `process-compose project
is-ready`, which checks if all the processes are started and ready with
the same logic as `--detach-on-success`. If the processes are not ready,
`process-compose` prints an error and fails with exit code 1.

If `--wait` is given, `process-compose` polls the processes on a
1-second delay until they're ready:

    $ PC_LOG_FILE=/dev/stdout ~/process-compose/bin/process-compose project is-ready --wait
    25-04-18 16:12:44.026 INF Process Compose v1.46.0
    25-04-18 16:12:44.026 WRN Get "http://unix/processes": dial unix /tmp/mwb-MjA3wz5a/pc: connect: no such file or directory
    25-04-18 16:12:45.033 WRN Processes are not ready:
    • mailpit (health is Unknown)
    • postgres (health is Unknown)
    • temporal (health is Unknown)
    25-04-18 16:12:46.038 WRN Processes are not ready:
    • postgres (health is Unknown)
    25-04-18 16:12:47.041 INF 3 processes are ready

In the future, the delay (and a timeout) should be configurable; see
issue #320.